Google Cloud Functions2/requirements.txt
google-cloud-storage
google-cloud-vision
google-api-core
grpcio
functions-framework==3.*
gcloud